export

+++ funasr/export/test/test_onnx_punc_vadrealtime.py
@@ -12,7 +12,7 @@ if __name__ == '__main__':
         return {'input': np.ones((1, text_length), dtype=np.int64),
                 'text_lengths': np.array([text_length,], dtype=np.int32),
                 'vad_mask': np.ones((1, 1, text_length, text_length), dtype=np.float32),
-                'sub_masks': np.tril(np.ones((text_length, text_length), dtype=np.float32))[None, None, :, :]
+                'sub_masks': np.tril(np.ones((text_length, text_length), dtype=np.float32))[None, None, :, :].astype(np.float32)
                 }
 
     def _run(feed_dict):
